requirements.txt
Django==1.6.2
distribute==0.6.34
djangorestframework==2.3.13
wsgiref==0.1.2
